refact: rank panel refactored to use react query
Add react-error-boundary module
