+++ package.json
@@ -21,6 +21,7 @@
     "axios": "^1.10.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-error-boundary": "^6.0.0",
     "react-hook-form": "^7.60.0",
     "react-router-dom": "^7.6.2",
     "react-toastify": "^11.0.5",
